add objectweb

+++ src/org/codelibs/sai/internal/objects/NativeFloat32Array.java
@@ -84,7 +84,7 @@ public final class NativeFloat32Array extends ArrayBufferView {
                 int.class, double.class).methodHandle();
 
         private Float32ArrayData(final FloatBuffer nb, final int start, final int end) {
-            super(((FloatBuffer) nb.position(start).limit(end)).slice(), end - start);
+            super((nb.position(start).limit(end)).slice(), end - start);
         }
 
         @Override
